[fix] Fix 23 bugs: FOK rollback, auth bypass, UB, overflow, XSS, OOM
Critical: FOK dry-run pre-check prevents book corruption, ptr::read_unaligned
for CMP payloads, X-User-Id fallback gated on dev mode only.

High: OrderSlot.sequence u16->u32, socialized loss i128 intermediate,
compression zone underflow/overflow guards, margin notional saturation.

Medium: pong payload clamped to 125 bytes, clock saturating_sub for NTP,
negative mark price fix, nightly API removal, WS max payload 1MB,
path traversal guard, XSS escaping in docs/WAL/risk endpoints.

Low: CID hex format, zero-latency truthiness fix, stress reports endpoint.

+++ rsx-book/src/book.rs
@@ -139,7 +139,7 @@ impl Orderbook {
         slot.next = NONE;
         slot.prev = NONE;
         self.sequence += 1;
-        slot.sequence = self.sequence as u16;
+        slot.sequence = self.sequence as u32;
 
         // Link into level
         let level =
